+++ tableauserverclient/server/__init__.py
@@ -2,12 +2,13 @@ from .request_factory import RequestFactory
 from .request_options import CSVRequestOptions, ImageRequestOptions, PDFRequestOptions, RequestOptions
 from .filter import Filter
 from .sort import Sort
-from .. import ConnectionItem, DatasourceItem, JobItem, BackgroundJobItem, \
+from .. import ConnectionItem, DatasourceItem, DatabaseItem, JobItem, BackgroundJobItem, \
     GroupItem, PaginationItem, ProjectItem, ScheduleItem, SiteItem, TableauAuth,\
-    UserItem, ViewItem, WorkbookItem, TaskItem, SubscriptionItem
+    UserItem, ViewItem, WorkbookItem, TableItem, TaskItem, SubscriptionItem, \
+    PermissionsRule, Permission, ColumnItem, FlowItem
 from .endpoint import Auth, Datasources, Endpoint, Groups, Projects, Schedules, \
-    Sites, Users, Views, Workbooks, Subscriptions, ServerResponseError, \
-    MissingRequiredFieldError
+    Sites, Tables, Users, Views, Workbooks, Subscriptions, ServerResponseError, \
+    MissingRequiredFieldError, Flows
 from .server import Server
 from .pager import Pager
 from .exceptions import NotSignedInError