+++ tableauserverclient/server/endpoint/__init__.py
@@ -1,13 +1,17 @@
 from .auth_endpoint import Auth
 from .datasources_endpoint import Datasources
+from .databases_endpoint import Databases
 from .endpoint import Endpoint
+from .flows_endpoint import Flows
 from .exceptions import ServerResponseError, MissingRequiredFieldError, ServerInfoEndpointNotFoundError
 from .groups_endpoint import Groups
 from .jobs_endpoint import Jobs
+from .metadata_endpoint import Metadata
 from .projects_endpoint import Projects
 from .schedules_endpoint import Schedules
 from .server_info_endpoint import ServerInfo
 from .sites_endpoint import Sites
+from .tables_endpoint import Tables
 from .tasks_endpoint import Tasks
 from .users_endpoint import Users
 from .views_endpoint import Views