+++ tableauserverclient/server/endpoint/server_info_endpoint.py
@@ -1,5 +1,5 @@
 from .endpoint import Endpoint, api
-from .exceptions import ServerResponseError, ServerInfoEndpointNotFoundError
+from .exceptions import ServerResponseError, ServerInfoEndpointNotFoundError, EndpointUnavailableError
 from ...models import ServerInfoItem
 import logging
 
@@ -19,6 +19,8 @@ class ServerInfo(Endpoint):
         except ServerResponseError as e:
             if e.code == "404003":
                 raise ServerInfoEndpointNotFoundError
+            if e.code == "404001":
+                raise EndpointUnavailableError
 
         server_info = ServerInfoItem.from_response(server_response.content, self.parent_srv.namespace)
         return server_info