+++ test/test_workbook.py
@@ -7,6 +7,10 @@ import xml.etree.ElementTree as ET
 from tableauserverclient.datetime_helpers import format_datetime
 from tableauserverclient.server.endpoint.exceptions import InternalServerError
 from tableauserverclient.server.request_factory import RequestFactory
+from tableauserverclient.models.permissions_item import PermissionsRule
+from tableauserverclient.models.user_item import UserItem
+from tableauserverclient.models.group_item import GroupItem
+
 from ._utils import asset
 
 TEST_ASSET_DIR = os.path.join(os.path.dirname(__file__), 'assets')
@@ -17,12 +21,14 @@ GET_EMPTY_XML = os.path.join(TEST_ASSET_DIR, 'workbook_get_empty.xml')
 GET_XML = os.path.join(TEST_ASSET_DIR, 'workbook_get.xml')
 POPULATE_CONNECTIONS_XML = os.path.join(TEST_ASSET_DIR, 'workbook_populate_connections.xml')
 POPULATE_PDF = os.path.join(TEST_ASSET_DIR, 'populate_pdf.pdf')
+POPULATE_PERMISSIONS_XML = os.path.join(TEST_ASSET_DIR, 'workbook_populate_permissions.xml')
 POPULATE_PREVIEW_IMAGE = os.path.join(TEST_ASSET_DIR, 'RESTAPISample Image.png')
 POPULATE_VIEWS_XML = os.path.join(TEST_ASSET_DIR, 'workbook_populate_views.xml')
 POPULATE_VIEWS_USAGE_XML = os.path.join(TEST_ASSET_DIR, 'workbook_populate_views_usage.xml')
 PUBLISH_XML = os.path.join(TEST_ASSET_DIR, 'workbook_publish.xml')
 PUBLISH_ASYNC_XML = os.path.join(TEST_ASSET_DIR, 'workbook_publish_async.xml')
 UPDATE_XML = os.path.join(TEST_ASSET_DIR, 'workbook_update.xml')
+UPDATE_PERMISSIONS = os.path.join(TEST_ASSET_DIR, 'workbook_update_permissions.xml')
 
 
 class WorkbookTests(unittest.TestCase):
@@ -269,6 +275,66 @@ class WorkbookTests(unittest.TestCase):
             self.assertEqual('dataengine', single_workbook.connections[0].connection_type)
             self.assertEqual('4506225a-0d32-4ab1-82d3-c24e85f7afba', single_workbook.connections[0].datasource_id)
             self.assertEqual('World Indicators', single_workbook.connections[0].datasource_name)
+
+    def test_populate_permissions(self):
+        with open(POPULATE_PERMISSIONS_XML, 'rb') as f:
+            response_xml = f.read().decode('utf-8')
+        with requests_mock.mock() as m:
+            m.get(self.baseurl + '/21778de4-b7b9-44bc-a599-1506a2639ace/permissions', text=response_xml)
+            single_workbook = TSC.WorkbookItem('test')
+            single_workbook._id = '21778de4-b7b9-44bc-a599-1506a2639ace'
+
+            self.server.workbooks.populate_permissions(single_workbook)
+            permissions = single_workbook.permissions
+
+            self.assertEqual(permissions[0].grantee.tag_name, 'group')
+            self.assertEqual(permissions[0].grantee.id, '5e5e1978-71fa-11e4-87dd-7382f5c437af')
+            self.assertDictEqual(permissions[0].capabilities, {
+                TSC.Permission.Capability.WebAuthoring: TSC.Permission.Mode.Allow,
+                TSC.Permission.Capability.Read: TSC.Permission.Mode.Allow,
+                TSC.Permission.Capability.Filter: TSC.Permission.Mode.Allow,
+                TSC.Permission.Capability.AddComment: TSC.Permission.Mode.Allow
+            })
+
+            self.assertEqual(permissions[1].grantee.tag_name, 'user')
+            self.assertEqual(permissions[1].grantee.id, '7c37ee24-c4b1-42b6-a154-eaeab7ee330a')
+            self.assertDictEqual(permissions[1].capabilities, {
+                TSC.Permission.Capability.ExportImage: TSC.Permission.Mode.Allow,
+                TSC.Permission.Capability.ShareView: TSC.Permission.Mode.Allow,
+                TSC.Permission.Capability.ExportData: TSC.Permission.Mode.Deny,
+                TSC.Permission.Capability.ViewComments: TSC.Permission.Mode.Deny
+            })
+
+    def test_add_permissions(self):
+        with open(UPDATE_PERMISSIONS, 'rb') as f:
+            response_xml = f.read().decode('utf-8')
+
+        single_workbook = TSC.WorkbookItem('test')
+        single_workbook._id = '21778de4-b7b9-44bc-a599-1506a2639ace'
+
+        bob = UserItem.as_reference("7c37ee24-c4b1-42b6-a154-eaeab7ee330a")
+        group_of_people = GroupItem.as_reference("5e5e1978-71fa-11e4-87dd-7382f5c437af")
+
+        new_permissions = [
+            PermissionsRule(bob, {'Write': 'Allow'}),
+            PermissionsRule(group_of_people, {'Read': 'Deny'})
+        ]
+
+        with requests_mock.mock() as m:
+            m.put(self.baseurl + "/21778de4-b7b9-44bc-a599-1506a2639ace/permissions", text=response_xml)
+            permissions = self.server.workbooks.update_permissions(single_workbook, new_permissions)
+
+        self.assertEqual(permissions[0].grantee.tag_name, 'group')
+        self.assertEqual(permissions[0].grantee.id, '5e5e1978-71fa-11e4-87dd-7382f5c437af')
+        self.assertDictEqual(permissions[0].capabilities, {
+            TSC.Permission.Capability.Read: TSC.Permission.Mode.Deny
+        })
+
+        self.assertEqual(permissions[1].grantee.tag_name, 'user')
+        self.assertEqual(permissions[1].grantee.id, '7c37ee24-c4b1-42b6-a154-eaeab7ee330a')
+        self.assertDictEqual(permissions[1].capabilities, {
+            TSC.Permission.Capability.Write: TSC.Permission.Mode.Allow
+        })
 
     def test_populate_connections_missing_id(self):
         single_workbook = TSC.WorkbookItem('test')